+++ src/main/java/com/github/castorm/kafka/connect/http/HttpSourceConnectorConfig.java
@@ -30,7 +30,9 @@ import com.github.castorm.kafka.connect.http.record.SchemedSourceRecordMapper;
 import com.github.castorm.kafka.connect.http.record.spi.SourceRecordMapper;
 import com.github.castorm.kafka.connect.http.request.offset.OffsetTemplateHttpRequestFactory;
 import com.github.castorm.kafka.connect.http.request.spi.HttpRequestFactory;
+import com.github.castorm.kafka.connect.http.response.PassthroughFilterFactory;
 import com.github.castorm.kafka.connect.http.response.jackson.JacksonHttpResponseParser;
+import com.github.castorm.kafka.connect.http.response.spi.HttpResponseFilterFactory;
 import com.github.castorm.kafka.connect.http.response.spi.HttpResponseParser;
 import lombok.Getter;
 import org.apache.kafka.common.config.AbstractConfig;
@@ -38,39 +40,50 @@ import org.apache.kafka.common.config.ConfigDef;
 
 import java.util.Map;
 
+import static com.github.castorm.kafka.connect.common.MapUtils.breakDownMap;
 import static org.apache.kafka.common.config.ConfigDef.Importance.HIGH;
+import static org.apache.kafka.common.config.ConfigDef.Importance.LOW;
 import static org.apache.kafka.common.config.ConfigDef.Type.CLASS;
+import static org.apache.kafka.common.config.ConfigDef.Type.STRING;
 
 @Getter
 class HttpSourceConnectorConfig extends AbstractConfig {
 
-    private static final String POLL_HOOKS = "http.poll.interceptor";
+    private static final String POLL_INTERCEPTOR = "http.poll.interceptor";
     private static final String CLIENT = "http.client";
     private static final String REQUEST_FACTORY = "http.request.factory";
     private static final String RESPONSE_PARSER = "http.response.parser";
+    private static final String RESPONSE_FILTER_FACTORY = "http.response.filter.factory";
     private static final String RECORD_MAPPER = "http.record.mapper";
+    private static final String OFFSET_INITIAL = "http.offset.initial";
 
     private final PollInterceptor pollInterceptor;
     private final HttpRequestFactory requestFactory;
     private final HttpClient client;
     private final HttpResponseParser responseParser;
+    private final HttpResponseFilterFactory responseFilterFactory;
     private final SourceRecordMapper recordMapper;
+    private final Map<String, String> initialOffset;
 
     HttpSourceConnectorConfig(Map<String, ?> originals) {
         super(config(), originals);
-        pollInterceptor = getConfiguredInstance(POLL_HOOKS, PollInterceptor.class);
+        pollInterceptor = getConfiguredInstance(POLL_INTERCEPTOR, PollInterceptor.class);
         requestFactory = getConfiguredInstance(REQUEST_FACTORY, HttpRequestFactory.class);
         client = getConfiguredInstance(CLIENT, HttpClient.class);
         responseParser = getConfiguredInstance(RESPONSE_PARSER, HttpResponseParser.class);
+        responseFilterFactory = getConfiguredInstance(RESPONSE_FILTER_FACTORY, HttpResponseFilterFactory.class);
         recordMapper = getConfiguredInstance(RECORD_MAPPER, SourceRecordMapper.class);
+        initialOffset = breakDownMap(getString(OFFSET_INITIAL));
     }
 
     public static ConfigDef config() {
         return new ConfigDef()
-                .define(POLL_HOOKS, CLASS, IntervalDelayPollInterceptor.class, HIGH, "Poll Interceptor Class")
+                .define(POLL_INTERCEPTOR, CLASS, IntervalDelayPollInterceptor.class, HIGH, "Poll Interceptor Class")
                 .define(CLIENT, CLASS, OkHttpClient.class, HIGH, "Request Client Class")
                 .define(REQUEST_FACTORY, CLASS, OffsetTemplateHttpRequestFactory.class, HIGH, "Request Factory Class")
                 .define(RESPONSE_PARSER, CLASS, JacksonHttpResponseParser.class, HIGH, "Response Parser Class")
-                .define(RECORD_MAPPER, CLASS, SchemedSourceRecordMapper.class, HIGH, "Record Mapper Class");
+                .define(RESPONSE_FILTER_FACTORY, CLASS, PassthroughFilterFactory.class, LOW, "Record Filter Factory Class")
+                .define(RECORD_MAPPER, CLASS, SchemedSourceRecordMapper.class, HIGH, "Record Mapper Class")
+                .define(OFFSET_INITIAL, STRING, "", HIGH, "Starting offset");
     }
 }