+++ src/main/java/com/github/castorm/kafka/connect/http/record/spi/SourceRecordMapper.java
@@ -26,7 +26,13 @@ import com.github.castorm.kafka.connect.http.model.HttpResponseItem;
 import org.apache.kafka.common.Configurable;
 import org.apache.kafka.connect.source.SourceRecord;
 
+import java.util.Map;
+
 public interface SourceRecordMapper extends Configurable {
 
     SourceRecord map(HttpResponseItem item);
+
+    default void configure(Map<String, ?> map) {
+        // Do nothing
+    }
 }