+++ src/main/java/com/github/castorm/kafka/connect/http/request/spi/HttpRequestFactory.java
@@ -23,15 +23,16 @@ package com.github.castorm.kafka.connect.http.request.spi;
  */
 
 import com.github.castorm.kafka.connect.http.model.HttpRequest;
+import com.github.castorm.kafka.connect.http.model.Offset;
 import org.apache.kafka.common.Configurable;
 
 import java.util.Map;
 
 public interface HttpRequestFactory extends Configurable {
 
-    void initializeOffset(Map<String, ?> offset);
+    HttpRequest createRequest(Offset offset);
 
-    void advanceOffset(Map<String, ?> offset);
-
-    HttpRequest createRequest();
+    default void configure(Map<String, ?> map) {
+        // Do nothing
+    }
 }