+++ src/main/java/com/github/castorm/kafka/connect/http/response/PassthroughFilterFactory.java
@@ -1,8 +1,8 @@
-package com.github.castorm.kafka.connect.http.model;
+package com.github.castorm.kafka.connect.http.response;
 
 /*-
  * #%L
- * kafka-connect-http-plugin
+ * Kafka Connect HTTP Plugin
  * %%
  * Copyright (C) 2020 CastorM
  * %%
@@ -22,18 +22,16 @@ package com.github.castorm.kafka.connect.http.model;
  * #L%
  */
 
-import lombok.Builder;
-import lombok.Value;
-import lombok.With;
+import com.github.castorm.kafka.connect.http.model.HttpResponseItem;
+import com.github.castorm.kafka.connect.http.model.Offset;
+import com.github.castorm.kafka.connect.http.response.spi.HttpResponseFilterFactory;
 
-@With
-@Value
-@Builder
-public class HttpResponseItem {
+import java.util.function.Predicate;
 
-    String key;
+public class PassthroughFilterFactory implements HttpResponseFilterFactory {
 
-    String value;
-
-    Offset offset;
+    @Override
+    public Predicate<HttpResponseItem> create(Offset offset) {
+        return __ -> true;
+    }
 }