+++ src/main/java/com/github/castorm/kafka/connect/http/response/spi/HttpResponseParser.java
@@ -27,8 +27,13 @@ import com.github.castorm.kafka.connect.http.model.HttpResponseItem;
 import org.apache.kafka.common.Configurable;
 
 import java.util.List;
+import java.util.Map;
 
 public interface HttpResponseParser extends Configurable {
 
     List<HttpResponseItem> parse(HttpResponse response);
+
+    default void configure(Map<String, ?> map) {
+        // Do nothing
+    }
 }