+++ src/test/java/com/github/castorm/kafka/connect/http/HttpSourceConnectorConfigTest.java
@@ -27,12 +27,14 @@ import com.github.castorm.kafka.connect.http.client.spi.HttpClient;
 import com.github.castorm.kafka.connect.http.model.HttpRequest;
 import com.github.castorm.kafka.connect.http.model.HttpResponse;
 import com.github.castorm.kafka.connect.http.model.HttpResponseItem;
+import com.github.castorm.kafka.connect.http.model.Offset;
 import com.github.castorm.kafka.connect.http.record.SchemedSourceRecordMapper;
 import com.github.castorm.kafka.connect.http.record.spi.SourceRecordMapper;
 import com.github.castorm.kafka.connect.http.request.offset.OffsetTemplateHttpRequestFactory;
 import com.github.castorm.kafka.connect.http.request.spi.HttpRequestFactory;
 import com.github.castorm.kafka.connect.http.response.jackson.JacksonHttpResponseParser;
 import com.github.castorm.kafka.connect.http.response.spi.HttpResponseParser;
+import com.google.common.collect.ImmutableMap;
 import org.apache.kafka.connect.source.SourceRecord;
 import org.junit.jupiter.api.Test;
 
@@ -42,6 +44,7 @@ import java.util.Map;
 
 import static com.github.castorm.kafka.connect.http.HttpSourceConnectorConfigTest.Fixture.config;
 import static com.github.castorm.kafka.connect.http.HttpSourceConnectorConfigTest.Fixture.configWithout;
+import static java.util.Collections.emptyMap;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class HttpSourceConnectorConfigTest {
@@ -86,26 +89,30 @@ class HttpSourceConnectorConfigTest {
         assertThat(config("http.record.mapper", TestRecordMapper.class.getName()).getRecordMapper()).isInstanceOf(TestRecordMapper.class);
     }
 
+    @Test
+    void whenNoInitialOffset_thenDefault() {
+        assertThat(configWithout("http.offset.initial").getInitialOffset()).isEqualTo(emptyMap());
+    }
+
+    @Test
+    void whenInitialOffset_thenInitialized() {
+        assertThat(config("http.offset.initial", "k=v").getInitialOffset()).isEqualTo(ImmutableMap.of("k", "v"));
+    }
+
     public static class TestHttpClient implements HttpClient {
         public HttpResponse execute(HttpRequest request) { return null; }
-        public void configure(Map<String, ?> map) {}
     }
 
     public static class TestRequestFactory implements HttpRequestFactory {
-        public void initializeOffset(Map<String, ?> offset) {}
-        public void advanceOffset(Map<String, ?> offset) {}
-        public HttpRequest createRequest() { return null; }
-        public void configure(Map<String, ?> map) {}
+        public HttpRequest createRequest(Offset offset) { return null; }
     }
 
     public static class TestResponseParser implements HttpResponseParser {
         public List<HttpResponseItem> parse(HttpResponse response) { return null; }
-        public void configure(Map<String, ?> map) {}
     }
 
     public static class TestRecordMapper implements SourceRecordMapper {
         public SourceRecord map(HttpResponseItem item) { return null; }
-        public void configure(Map<String, ?> map) {}
     }
 
     interface Fixture {